Reorganizing and removing bot dependency in db file
