+++ src/db/users.ts
@@ -1,6 +1,5 @@
 import prisma from '.';
 import type { User } from '.prisma/client';
-import { client } from '../../bot';
 
 export const getUserById = async (id: string): Promise<User | null> => {
   const user = await prisma.user.findUnique({
@@ -11,27 +10,13 @@ export const getUserById = async (id: string): Promise<User | null> => {
   return user;
 };
 
-export const createUser = async (id: string, username: string = '') => {
+export const createUser = async (id: string, username: string) => {
   return await prisma.user.create({
     data: {
       id,
       username,
     },
   });
-};
-
-export const getOrCreateUser = async (id: string, username?: string) => {
-  const existingUser = await getUserById(id);
-  if (existingUser) {
-    return existingUser;
-  }
-  if (!username) {
-    const user = await client.users.fetch(id);
-    username = user.username;
-  }
-  const createdUser = await createUser(id, username);
-  console.info({ createdUser });
-  return createdUser;
 };
 
 export const deleteUser = async (id: string): Promise<User | null> => {
