+++ src/utils/interfaces.ts
@@ -1,0 +1,8 @@
+export interface Leader {
+  username: string;
+  totalPoints: number;
+  learnPoints: number;
+  buildPoints: number;
+  teachPoints: number;
+  id: string;
+}
